Revert "!24935 【无障碍支持UI Extension查询】无障碍id由int32_t变更为int64_t"
This reverts commit a295570e6a17480ea520c04c734875ae5232721a, reversing
changes made to f8589be73f50942d1a86e4002fcdd9a883cd7f20.

+++ frameworks/core/components_ng/base/ui_node.cpp
@@ -33,7 +33,7 @@
 
 namespace OHOS::Ace::NG {
 
-thread_local int64_t UINode::currentAccessibilityId_ = 0;
+thread_local int32_t UINode::currentAccessibilityId_ = 0;
 
 UINode::UINode(const std::string& tag, int32_t nodeId, bool isRoot)
     : tag_(tag), nodeId_(nodeId), accessibilityId_(currentAccessibilityId_++), isRoot_(isRoot)
@@ -1004,7 +1004,7 @@ std::string UINode::GetCurrentCustomNodeInfo()
     return extraInfo;
 }
 
-int64_t UINode::GenerateAccessibilityId()
+int32_t UINode::GenerateAccessibilityId()
 {
     return currentAccessibilityId_++;
 }
