+++ frameworks/core/components_ng/pattern/ui_extension/session_wrapper_impl.cpp
@@ -88,8 +88,7 @@ public:
             TaskExecutor::TaskType::UI);
     }
 
-    void OnAccessibilityEvent(
-        const Accessibility::AccessibilityEventInfo& info, int64_t uiExtensionOffset) override
+    void OnAccessibilityEvent(const Accessibility::AccessibilityEventInfo& info, int32_t uiExtensionOffset) override
     {
         ContainerScope scope(instanceId_);
         auto pipeline = PipelineBase::GetCurrentContext();
@@ -387,20 +386,20 @@ void SessionWrapperImpl::NotifyConfigurationUpdate() {}
 
 /************************************************ Begin: The interface about the accessibility ************************/
 bool SessionWrapperImpl::TransferExecuteAction(
-    int64_t elementId, const std::map<std::string, std::string>& actionArguments, int32_t action, int64_t offset)
+    int32_t elementId, const std::map<std::string, std::string>& actionArguments, int32_t action, int32_t offset)
 {
     CHECK_NULL_RETURN(session_, false);
     return OHOS::Rosen::WSError::WS_OK == session_->TransferExecuteAction(elementId, actionArguments, action, offset);
 }
 
 void SessionWrapperImpl::SearchExtensionElementInfoByAccessibilityId(
-    int64_t elementId, int32_t mode, int64_t baseParent, std::list<Accessibility::AccessibilityElementInfo>& output)
+    int32_t elementId, int32_t mode, int32_t baseParent, std::list<Accessibility::AccessibilityElementInfo>& output)
 {
     CHECK_NULL_VOID(session_);
     session_->TransferSearchElementInfo(elementId, mode, baseParent, output);
 }
 
-void SessionWrapperImpl::SearchElementInfosByText(int64_t elementId, const std::string& text, int64_t baseParent,
+void SessionWrapperImpl::SearchElementInfosByText(int32_t elementId, const std::string& text, int32_t baseParent,
     std::list<Accessibility::AccessibilityElementInfo>& output)
 {
     CHECK_NULL_VOID(session_);
@@ -408,14 +407,14 @@ void SessionWrapperImpl::SearchElementInfosByText(int64_t elementId, const std::
 }
 
 void SessionWrapperImpl::FindFocusedElementInfo(
-    int64_t elementId, int32_t focusType, int64_t baseParent, Accessibility::AccessibilityElementInfo& output)
+    int32_t elementId, int32_t focusType, int32_t baseParent, Accessibility::AccessibilityElementInfo& output)
 {
     CHECK_NULL_VOID(session_);
     session_->TransferFindFocusedElementInfo(elementId, focusType, baseParent, output);
 }
 
 void SessionWrapperImpl::FocusMoveSearch(
-    int64_t elementId, int32_t direction, int64_t baseParent, Accessibility::AccessibilityElementInfo& output)
+    int32_t elementId, int32_t direction, int32_t baseParent, Accessibility::AccessibilityElementInfo& output)
 {
     CHECK_NULL_VOID(session_);
     session_->TransferFocusMoveSearch(elementId, direction, baseParent, output);
